requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
ray
omegaconf
transformers
vllm
wandb
swanlab
pylatexenc
vertexai
sentence_transformers
packaging>=20.0
uvicorn
fastapi
